update SVM code
